requirements.txt
flask
pytesseract
Pillow

